+++ operator/internal/manifests/gateway.go
@@ -99,6 +99,10 @@ func BuildGateway(opts Options) ([]client.Object, error) {
 		}
 	}
 
+	if err := configureReplication(&dpl.Spec.Template, opts.Stack.Replication, LabelGatewayComponent, opts.Name); err != nil {
+		return nil, err
+	}
+
 	return objs, nil
 }
 
@@ -107,9 +111,8 @@ func NewGatewayDeployment(opts Options, sha1C string) *appsv1.Deployment {
 	l := ComponentLabels(LabelGatewayComponent, opts.Name)
 	a := commonAnnotations(sha1C, opts.CertRotationRequiredAt)
 	podSpec := corev1.PodSpec{
-		ServiceAccountName:        GatewayName(opts.Name),
-		Affinity:                  configureAffinity(LabelGatewayComponent, opts.Name, opts.Gates.DefaultNodeAffinity, opts.Stack.Template.Gateway),
-		TopologySpreadConstraints: defaultTopologySpreadConstraints(LabelGatewayComponent, opts.Name),
+		ServiceAccountName: GatewayName(opts.Name),
+		Affinity:           configureAffinity(LabelGatewayComponent, opts.Name, opts.Gates.DefaultNodeAffinity, opts.Stack.Template.Gateway),
 		Volumes: []corev1.Volume{
 			{
 				Name: "rbac",