+++ operator/internal/manifests/var.go
@@ -123,13 +123,8 @@ const (
 var (
 	defaultTimeoutConfig = calculateHTTPTimeouts(lokiDefaultQueryTimeout)
 
-	defaultConfigMapMode      = int32(420)
-	volumeFileSystemMode      = corev1.PersistentVolumeFilesystem
-	podAntiAffinityComponents = map[string]struct{}{
-		LabelIngesterComponent:      {},
-		LabelRulerComponent:         {},
-		LabelQueryFrontendComponent: {},
-	}
+	defaultConfigMapMode = int32(420)
+	volumeFileSystemMode = corev1.PersistentVolumeFilesystem
 )
 
 func commonAnnotations(configHash, rotationRequiredAt string) map[string]string {
@@ -148,24 +143,11 @@ func commonLabels(stackName string) map[string]string {
 	}
 }
 
-func componentInstaceLabels(component string, stackName string) map[string]string {
+func componentInstanceLabels(component string, stackName string) map[string]string {
 	return map[string]string{
 		kubernetesInstanceLabel:  stackName,
 		kubernetesComponentLabel: component,
 	}
-}
-
-// defaultTopologySpreadConstraints returns a topology spread contraint that will
-// instruct the scheduler to try and schedule pods from the same component in different nodes
-func defaultTopologySpreadConstraints(component string, stackName string) []corev1.TopologySpreadConstraint {
-	return []corev1.TopologySpreadConstraint{{
-		MaxSkew:     1,
-		TopologyKey: kubernetesNodeHostnameLabel,
-		LabelSelector: &metav1.LabelSelector{
-			MatchLabels: componentInstaceLabels(component, stackName),
-		},
-		WhenUnsatisfiable: corev1.ScheduleAnyway,
-	}}
 }
 
 func serviceAnnotations(serviceName string, enableSigningService bool) map[string]string {
@@ -174,28 +156,6 @@ func serviceAnnotations(serviceName string, enableSigningService bool) map[strin
 		annotations[openshift.ServingCertKey] = serviceName
 	}
 	return annotations
-}
-
-func topologySpreadConstraints(spec lokiv1.ReplicationSpec, component string, stackName string) []corev1.TopologySpreadConstraint {
-	var tsc []corev1.TopologySpreadConstraint
-	if len(spec.Zones) > 0 {
-		tsc = make([]corev1.TopologySpreadConstraint, len(spec.Zones))
-		for i, z := range spec.Zones {
-			tsc[i] = corev1.TopologySpreadConstraint{
-				MaxSkew:           int32(z.MaxSkew),
-				TopologyKey:       z.TopologyKey,
-				WhenUnsatisfiable: corev1.DoNotSchedule,
-				LabelSelector: &metav1.LabelSelector{
-					MatchLabels: map[string]string{
-						kubernetesComponentLabel: component,
-						kubernetesInstanceLabel:  stackName,
-					},
-				},
-			}
-		}
-	}
-
-	return tsc
 }
 
 // ComponentLabels is a list of all commonLabels including the app.kubernetes.io/component:<component> label
@@ -532,10 +492,6 @@ func configureAffinity(componentLabel, stackName string, enableNodeAffinity bool
 	if cSpec.PodAntiAffinity != nil {
 		affinity.PodAntiAffinity = cSpec.PodAntiAffinity
 	}
-
-	if affinity.NodeAffinity == nil && affinity.PodAntiAffinity == nil {
-		return nil
-	}
 	return affinity
 }
 
@@ -567,18 +523,13 @@ func defaultNodeAffinity(enableNodeAffinity bool) *corev1.NodeAffinity {
 // defaultPodAntiAffinity for components in podAntiAffinityComponents will
 // configure pods, of a LokiStack, to preferably not run on the same node
 func defaultPodAntiAffinity(componentLabel, stackName string) *corev1.PodAntiAffinity {
-	_, enablePodAntiAffinity := podAntiAffinityComponents[componentLabel]
-	if !enablePodAntiAffinity {
-		return nil
-	}
-
 	return &corev1.PodAntiAffinity{
 		PreferredDuringSchedulingIgnoredDuringExecution: []corev1.WeightedPodAffinityTerm{
 			{
 				Weight: 100,
 				PodAffinityTerm: corev1.PodAffinityTerm{
 					LabelSelector: &metav1.LabelSelector{
-						MatchLabels: componentInstaceLabels(componentLabel, stackName),
+						MatchLabels: componentInstanceLabels(componentLabel, stackName),
 					},
 					TopologyKey: kubernetesNodeHostnameLabel,
 				},